:recycle: refactor: 잘못된 코드 위치 변경
